sd: send over sign key
Previously, we were trying to use a box keypair to sign messages,
which does not work because signing requires ed25519 keys while
encryption requires curve25519 keys. As such we were unable to
verify the signature received on the other side.

Instead, use an ed25519 keypair and send its public key to the
other side. While the public key cannot be used for
encryption/decryption, we can derive the curve25519 public key
from it and store it for later usage.

Instead of deriving the curve25519 from the ed25519 key, we
might instead consider using two different key pairs for signing
and authenticated encryption.