sd: test: fix setup/teardown logic
Previously we have been setting up the setup and teardown logic
to run before and after a whole group. The intention was to
instead run these functions before and after each unit test,
though.

Fix this by introducing a new macro which automatically sets the
setup and teardown function for each unit test.