sd: work out sign/encrypt key differences
When performing the initial hand shake we send over the long term
signing key, which is from there on used for both signatures and
encryption of the messages. Using the signature key as encryption
key is considered insecure though, as it provides an encryption
oracle to the adversary.

Rework the hand shake to always generate a new key used for
encryption. We will then send over our long term public signature
key and the newly generated encryption key, along with a
signature of the encryption key done with the long term signature
key. Like this the receiver can assert the encryption key is
actually valid for the given signature key.

This also fixes problems with generating nonces. As the signature
is always done over random data due to the session key being
generated at random we do not need to mind padding the message
with additional random data to avoid duplicated cipher texts. And
due to the encryption keys being generated for each new session
we can always start counting the nonces from 0.